% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortchar.R
\name{setcolor}
\alias{setcolor}
\title{set colors by giving length}
\usage{
setcolor(x)
}
\arguments{
\item{x}{number indicate length of color you want have}
}
\description{
set colors by giving length
}
\author{
Kai Guo
}
