% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortchar.R
\name{sortchar}
\alias{sortchar}
\title{sort character by the number inside}
\usage{
sortchar(x, order = FALSE)
}
\arguments{
\item{x}{vector of character}

\item{order}{use order or not}
}
\description{
sort character by the number inside
}
\author{
Kai Guo
}
